ostree-prepare-root: move /run mount point to sysroot/run
Move /run mount point from initramfs to real root filesystem

This change relocates the /run mount point from initramfs (/run) to the real root filesystem (sysroot/run). This ensures that systemd does not remount it during switchroot if it detects an existing mount point. Consequently, all contents of /run will be transferred to the real root filesystem.

This adjustment resolves the issue of mounting sysroot as read-only. During the execution of ostree-remount.service, the system checks /run/ostree-booted to determine if sysroot should be mounted as read-only.