Add support for bootloader-plugin
Can be used to provide hooks into a sysroot deploy for tools like
bootupd to perform bootloader swaps.

Happy to do this a different way, but this is how we've been doing
it on our debian-based system. I understand bootupd is supposed to
work a different way. We do it this way because grub-mkconfig will
generate a broken config for the older version of grub.