tests: account for different locales in remote-gpg-list-keys test
During test preparation in libtest-core.sh either C or en_US locale
is set for the test execution.

The remote-gpg-list-keys compares outputs that contain locale-dependent
dates, but the current test case expects the result to be in C locale.

However in case C.utf-8 locale is not available, but en_US.utf-8 locale
is available, then the latter will be used, and the testcase will fail,
because it expects dates to be in this format:

Tue 17 Mar 2015 02:00:32 PM UTC

However with US locale it receives the date like this:

Tue Mar 17 2015 02:00:32 PM UTC

This change converts these dates to the current locale before comparing
the expected gpg key content with the received one.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>